feat: use Coroutine for translateValue
